+++ Extension/src/background/api/filters/common.ts
@@ -33,7 +33,7 @@ import {
 import { network } from '../network';
 
 import { CustomFilterApi } from './custom';
-import { FiltersApi } from './main';
+import { type FilterMetadata, FiltersApi } from './main';
 import { type FilterUpdateOptions } from './update';
 import { FilterParser } from './parser';
 
@@ -81,6 +81,19 @@ export class CommonFilterApi {
             && filterId !== AntiBannerFiltersId.UserFilterId
             && filterId !== AntiBannerFiltersId.AllowlistFilterId
             && filterId !== AntiBannerFiltersId.QuickFixesFilterId;
+    }
+
+    /**
+     * Checks whether the filter is a regular filter.
+     *
+     * It is needed only for proper types checking instead of type castings.
+     *
+     * @param filter Filter metadata.
+     *
+     * @returns True if filter is a regular filter, false otherwise.
+     */
+    public static isRegularFilterMetadata(filter: FilterMetadata): filter is RegularFilterMetadata {
+        return CommonFilterApi.isCommonFilter(filter.filterId);
     }
 
     /**
@@ -139,6 +152,8 @@ export class CommonFilterApi {
      * dynamically.
      *
      * @returns Filter metadata — {@link RegularFilterMetadata}.
+     *
+     * @throws Error if filter is deprecated and should not be used.
      */
     public static async loadFilterRulesFromBackend(
         filterUpdateOptions: FilterUpdateOptions,
@@ -186,7 +201,13 @@ export class CommonFilterApi {
             expires,
             timeUpdated,
             diffPath,
+            deprecated,
+            filterId,
         } = filterMetadata;
+
+        if (deprecated) {
+            throw new Error(`Filter with id ${filterId} is deprecated and shall not be used`);
+        }
 
         const filterVersion = filterVersionStorage.get(filterUpdateOptions.filterId);