+++ tests/src/background/api/update.test.ts
@@ -116,20 +116,6 @@ describe('Update Api (without indexedDB)', () => {
             dateNowSpy.mockRestore();
         });
 
-        const v0 = getStorageFixturesV0();
-        const v1 = getStorageFixturesV1();
-        const v2 = getStorageFixturesV2(expires);
-        const v3 = getStorageFixturesV3(expires);
-        const v4 = getStorageFixturesV4(expires);
-        const v5 = getStorageFixturesV5(expires);
-        const v6 = getStorageFixturesV6(expires);
-        const v7 = getStorageFixturesV7(expires);
-        const v8 = getStorageFixturesV8(expires);
-        const v9 = getStorageFixturesV9(expires);
-        const v10 = getStorageFixturesV10(expires);
-        const v11 = getStorageFixturesV11(expires);
-        const v12 = getStorageFixturesV12(expires);
-
         const getCases = (from: StorageData[], to: StorageData[]) => {
             const cases = [];
 
@@ -155,18 +141,35 @@ describe('Update Api (without indexedDB)', () => {
             expect(deepJsonParse(settings)).toStrictEqual(deepJsonParse(data.to));
         };
 
-        it.each(getCases(v0, v12))('should update from v0 to v12', runCase);
-        it.each(getCases(v1, v12))('should update from v1 to v12', runCase);
-        it.each(getCases(v2, v12))('should update from v2 to v12', runCase);
-        it.each(getCases(v3, v12))('should update from v3 to v12', runCase);
-        it.each(getCases(v4, v12))('should update from v4 to v12', runCase);
-        it.each(getCases(v5, v12))('should update from v5 to v12', runCase);
-        it.each(getCases(v6, v12))('should update from v6 to v12', runCase);
-        it.each(getCases(v7, v12))('should update from v7 to v12', runCase);
-        it.each(getCases(v8, v12))('should update from v8 to v12', runCase);
-        it.each(getCases(v9, v12))('should update from v9 to v12', runCase);
-        it.each(getCases(v10, v12))('should update from v10 to v12', runCase);
-        it.each(getCases(v11, v12))('should update from v11 to v12', runCase);
+        const migrationCasesData = {
+            v0: getStorageFixturesV0(),
+            v1: getStorageFixturesV1(),
+            v2: getStorageFixturesV2(expires),
+            v3: getStorageFixturesV3(expires),
+            v4: getStorageFixturesV4(expires),
+            v5: getStorageFixturesV5(expires),
+            v6: getStorageFixturesV6(expires),
+            v7: getStorageFixturesV7(expires),
+            v8: getStorageFixturesV8(expires),
+            v9: getStorageFixturesV9(expires),
+            v10: getStorageFixturesV10(expires),
+            v11: getStorageFixturesV11(expires),
+            v12: getStorageFixturesV12(expires),
+        };
+
+        const targetVersion = Object.keys(migrationCasesData).pop() as keyof typeof migrationCasesData;
+
+        // Iterate over all versions except the last one
+        // e.g. if we have v0, v1, v2, v3, only v0->v3, v1->v3, v2->v3 will be tested
+        // but not v3->v3
+        Object.entries(migrationCasesData).forEach(([version, data]) => {
+            if (version === targetVersion) {
+                return;
+            }
+
+            const cases = getCases(data, migrationCasesData[targetVersion]);
+            it.each(cases)(`should update from ${version} to ${targetVersion}`, runCase);
+        });
 
         // Separate test for migration from V3 storage, because after this
         // version we moved from localStorage to hybridStorage.